+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added billsafe btn tansaction number
 - Added product first import unit name
 - Added product variation set default (standard) type
+- Added crossselling group support (sw_related, sw_similar)
 - Fixed wrong customer type hint in customer mapper class
 - Fixed missing crossselling linking